refactor: rename path resolver
